moving validation functions to separate file
